An interim check-in while I'm working to reorganize how the book contents are parsed so the DAISY 202 books can be read and navigated correctly. Currently the DaisyParser simply extracts the headers and 'a' elements as independent elements, sadly this results in audio being played multiple times when a book has several references to the same smil file e.g. for page numbers. The parser needs to track the parent element for each 'a' element e.g. some may belong to a <h1> element, others to 'page-normal' i.e. to a page number.
This change includes some new tests, revised method names in DaisyParser, and minor fiddles to a couple of other classes. There's more work to do before the application is worth building and releasing. Unless you're keen to help restructure the code, or work on the parsing, you can safely ignore this revision :)

Julian