This is a rough first cut of code to extract the text contents from text-based DAISY 2.02 books.
I'm not yet convinced about the data structures which differ if <par> tags wrap <audio> or <text> contents. Also, the only implementation is currently in ProcessExternalSmilFile. 

However, I'm checking this code in for posterity and so we can track the changes we make to the code and to the interface(s). I hope you'll also see me tidying up the code in future releases :)