Relatively major changes to enable 'parts' to be processed. We're getting closer to a working Android app :)
There are still some rough edges to the code, and the Android app fails at runtime because of a problem of Daisy202Section.Builder().build() not being visible even though ant and eclipse seem to think build() exists and is visible to callers. Ho hum, I'll find out what's going wrong and fix it.